add inline srat mode

+++ handlers/srat.py
@@ -1,5 +1,8 @@
+from typing import Optional
+
 from aiogram import types, Router, F
 from aiogram.enums import ChatType
+from aiogram.types import InlineQueryResultArticle, InlineQuery, InputTextMessageContent, ChosenInlineResult
 
 from db.ToiletSessions import SretSession
 from db.User import User
@@ -9,33 +12,35 @@ from utils import send_srat_notification
 router = Router()
 
 
-@router.message(F.text.startswith('Я'), F.chat.type == ChatType.PRIVATE)
-async def send_srat(message: types.Message, user: User):
-    if message.text == SretActions.SRET.value:
-        sret = 1
-
-    elif message.text == SretActions.END.value:
-        sret = 0
-
-    elif message.text == SretActions.PERNUL.value:
-        sret = 3
-
-    elif message.text == SretActions.DRISHET.value:
-        sret = 2
-
-    else:
-        return
-
+async def verify_action(user: User, sret: int, message: Optional[types.Message] = None):
     if sret in send_srat_notification.must_sret:
         if not await SretSession.filter(user=user, end=None).exists():
-            await message.reply('Ты что заканчивать захотел? Ты даже не срешь!')
-            return
+            if message is not None:
+                await message.reply('Ты что заканчивать захотел? Ты даже не срешь!')
+            return False
 
     if sret in send_srat_notification.must_not_sret:
         if await SretSession.filter(user=user, end=None).exists():
-            await message.reply('Ты прошлое свое сранье не закончил, а уже новое начинаешь?\n'
-                                'Нет уж. Будь добр, раз начал - закончи.')
-            return
+            if message is not None:
+                await message.reply('Ты прошлое свое сранье не закончил, а уже новое начинаешь?\n'
+                                    'Нет уж. Будь добр, раз начал - закончи.')
+            return False
+
+    return True
+
+
+@router.message(F.text.startswith('Я'), F.chat.type == ChatType.PRIVATE)
+async def send_srat(message: types.Message, user: User):
+    sret = None
+    for el in SretActions:
+        if el.value[1] == message.text:
+            sret = el.value[0]
+            break
+    if sret is None:
+        return
+
+    if not await verify_action(user, sret, message):
+        return
 
     await message.delete()
     await send_srat_notification.send(user, sret)
@@ -47,3 +52,31 @@ async def cancel_srat(callback: types.CallbackQuery):
 
     await callback.answer('Отменили автозавершение.')
     await callback.message.edit_reply_markup()
+
+
+@router.inline_query()
+async def get_sret_actions(inline_query: InlineQuery, user: User):
+    res = []
+    exists = await SretSession.filter(user=user, end=None).exists()
+    for el in SretActions:
+        sret = int(el.value[0])
+        in_must = sret in send_srat_notification.must_sret
+        inn_must = sret in send_srat_notification.must_not_sret
+        if in_must and exists or \
+                inn_must and not exists or (not in_must and not inn_must):
+            res.append(InlineQueryResultArticle(
+                id=str(el.value[0]),
+                title=el.value[1],
+                input_message_content=InputTextMessageContent(
+                    message_text=send_srat_notification.get_message_text(user, el.value[0])
+                )
+            ))
+
+    await inline_query.answer(res, is_personal=True, cache_time=0)
+
+
+@router.chosen_inline_result()
+async def send_srat_inline(chosen_result: ChosenInlineResult, user: User):
+    sret = int(chosen_result.result_id)
+    if await verify_action(user, sret):
+        await send_srat_notification.send(user, sret)
